0508/0508-149/requirements.txt
PyQt5>=5.15.0
send2trash>=1.8.0
Pillow>=8.0.0
